Allow covered content to optionally *not* convert to inline
Consider the case of `cover-with-rect(block(...))`. This should explicitly be in its own block, but adding a box deterministically inside `cover` will promote this to inline content. Providing an optional argument allows block-level content to stay that way if desired. "inline" is still a sensible default, since paragraph breaks around a box will allow it to be its own block anyway.